refactor: Доделать дз
